Split out generate-jsx utility
